Move getLocation, getZipLocation and getRootLocation
