+++ storage/storage.go
@@ -8,6 +8,8 @@ import (
 	"io"
 	"io/fs"
 	"os"
+	"path/filepath"
+	"strconv"
 )
 
 // Storage holds the methods to interact with a Cloud provider storage.
@@ -94,4 +96,27 @@ func UploadZip(ctx context.Context, resource Resource, file *os.File, fn WalkDir
 		return err
 	}
 	return nil
+}
+
+// getLocation returns the location of a Resource relative to the base location.
+//
+//	If path is not empty, it will append the given path to the resulting location of the resource.
+func getLocation(base string, r Resource, path string) string {
+	location := filepath.Join(base, r.GetOwner(), r.GetUUID(), strconv.FormatUint(r.GetVersion(), 10))
+	if len(path) > 0 {
+		location = filepath.Join(location, path)
+	}
+	return location
+}
+
+// getZipLocation returns the location of the zip file associated to a Resource relative to the base location.
+func getZipLocation(base string, r Resource) string {
+	filename := fmt.Sprintf("%d.zip", r.GetVersion())
+	return filepath.Join(base, r.GetOwner(), r.GetUUID(), ".zips", filename)
+}
+
+// getRootLocation returns the absolute location of where all the versions of the given uuid and the given kind will be
+// uploaded for the given owner.
+func getRootLocation(base string, owner string, uuid string) string {
+	return filepath.Join(base, owner, uuid)
 }
